Added TopologyManager to abstract replication topology configuration. Unified logging.
