+++ pkg/controller/replication/controller.go
@@ -53,7 +53,7 @@ type ReplicationReconciler struct {
 	recorder            events.EventRecorder
 	builder             *builder.Builder
 	env                 *environment.OperatorEnv
-	replConfigClient    *ReplicationConfigClient
+	topologyManager     *TopologyManager
 	refResolver         *refresolver.RefResolver
 	secretReconciler    *secret.SecretReconciler
 	configMapreconciler *configmap.ConfigMapReconciler
@@ -61,13 +61,13 @@ type ReplicationReconciler struct {
 }
 
 func NewReplicationReconciler(client client.Client, recorder events.EventRecorder, builder *builder.Builder, env *environment.OperatorEnv,
-	replConfigClient *ReplicationConfigClient, opts ...Option) (*ReplicationReconciler, error) {
+	topologyManager *TopologyManager, opts ...Option) (*ReplicationReconciler, error) {
 	r := &ReplicationReconciler{
-		Client:           client,
-		recorder:         recorder,
-		builder:          builder,
-		env:              env,
-		replConfigClient: replConfigClient,
+		Client:          client,
+		recorder:        recorder,
+		builder:         builder,
+		env:             env,
+		topologyManager: topologyManager,
 	}
 	for _, setOpt := range opts {
 		setOpt(r)
@@ -234,6 +234,7 @@ func (r *ReplicationReconciler) ReconcileReplicationInPod(ctx context.Context, r
 	replStatus := ptr.Deref(req.mariadb.Status.Replication, mariadbv1alpha1.ReplicationStatus{})
 	replRoles := replStatus.Roles
 	pod := statefulset.PodName(req.mariadb.ObjectMeta, podIndex)
+	topology := r.topologyManager.TopologyForMariaDB(req.mariadb, logger.WithValues("pod", pod))
 
 	if primaryPodIndex == podIndex {
 		if role, ok := replRoles[pod]; ok && role == mariadbv1alpha1.ReplicationRolePrimary {
@@ -244,8 +245,8 @@ func (r *ReplicationReconciler) ReconcileReplicationInPod(ctx context.Context, r
 			logger.V(1).Info("error getting current primary client", "err", err, "pod", pod)
 			return ctrl.Result{RequeueAfter: 5 * time.Second}, nil
 		}
-		logger.Info("Configuring primary", "pod", pod)
-		if err := r.replConfigClient.ConfigurePrimary(ctx, req.mariadb, client); err != nil {
+
+		if err := topology.ConfigurePrimary(ctx, client); err != nil {
 			return ctrl.Result{}, fmt.Errorf("error configuring replica: %v", err)
 		}
 		return ctrl.Result{}, nil
@@ -263,13 +264,12 @@ func (r *ReplicationReconciler) ReconcileReplicationInPod(ctx context.Context, r
 		logger.V(1).Info("error getting replica client", "err", err, "pod", pod)
 		return ctrl.Result{RequeueAfter: 5 * time.Second}, nil
 	}
-	logger.Info("Configuring replica", "pod", pod)
 
 	replicaOpts, err := r.getReplicaOpts(ctx, req, pod, podIndex, logger, reconcilePodOpts...)
 	if err != nil {
 		return ctrl.Result{}, fmt.Errorf("error getting replica opts: %v", err)
 	}
-	if err := r.replConfigClient.ConfigureReplica(ctx, req.mariadb, client, primaryPodIndex, replicaOpts...); err != nil {
+	if err := topology.ConfigureReplica(ctx, client, primaryPodIndex, replicaOpts...); err != nil {
 		return ctrl.Result{}, fmt.Errorf("error configuring replica: %v", err)
 	}
 	return ctrl.Result{}, nil
