--- pkg/multicluster/multicluster.go
@@ -1,3 +1,0 @@
-package multicluster
-
-var ReplicaConnectionName = "replica"
